Automatically switch between unpkg and jsdelivr
The currently selected package registry is stored in a
compute-persistent KV store called 'rum-collector-kvstore' under key
'PackageRegistry'. If the currently used registry reports errors, the
value is changed to the other registry in the store and a retry is done.

This commit contains the logic for fastly only, cloudflare is TODO.