feat(security): add X-Frame-Options header to responses
This update introduces the 'X-Frame-Options' header set to 'DENY' across various response functions to enhance security by preventing clickjacking attacks. The header is added to cleanupHeaders, cleanupResponse, prohibitDirectoryRequest, respondError, respondInfo, respondCORS, and robots.txt responses. Additionally, tests have been updated to verify the presence of this header in responses.
As none of the resources served by this service are of type text/html, there is no actual risk of click-jacking, so the addition of this header only serves the purpose of reducing the number of false-positive reports generated by overly simplistic security scanning software and their ill-configured reporting rules.